commander: cleanup some repetitive code
There were 3 call-sites where the exact same code was being called
for the exact same reason. Converted these call-sites to use an
array_size() function instead.

Also moved the calculation of the maximum number of CAPs allowed based
on player count to a function instead of it hanging out in the middle
of the commander function.

Signed-off-by: Jonathan Toppins <jtoppins@users.sourceforge.net>