commander: consider loss of assets when spawning assets
Modifies commander to consider in-game assets as part of its calculation
for what kind and how many CAPs and other assets to spawn.

This is an initial pass, refinement and tweaking is expected.
Example: currently the maximum number of CAPs and BAI are not limited
based on the logistics stat.

v2:
  added notes to why particular utility functions were used
  removed some todos
  simplified some functions based on reviews

Signed-off-by: Jonathan Toppins <jtoppins@users.sourceforge.net>